Create integration with Amazom Lex
we use Kommunicate to integrate the chatbot with the website we already built. It requires Access Key and Secret Access Key inorder to access the chatbot we have built in Amazon Lex. We can also enable the human entry feature which allows us to customise the chatbot responses whenever we’re active.